sks ta: generate a SKS return code when possible
For the SKS to provide client a meaningful retrun code, the GPD TEE
command invocation return code is forced to TEE_SUCCESS and client
is expected to read to effective SKS return code (which can notify
a success or a failure) from the SK TA API.

The first memref argument is used to store this SKS statuc info
when defined as an input/output memref by the client. In such case,
the output memref is a 32bit buffer that stores the SKS return code.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>